fixes for review issues in #5
