+++ app/javascript/src/components/Common/Sidebar/constants.js
@@ -5,7 +5,7 @@ export const APP_NAME = "Wheel";
 export const PASSWORD_PATH = "/my/password/edit";
 export const PROFILE_PATH = "/my/profile";
 export const LOGOUT_PATH = "/logout";
-export const USR_IMG_TMP =
+export const TEMP_USER_PROFILE_IMAGE_URL =
   "https://randomuser.me/api/portraits/med/women/54.jpg";
 
 export const SIDENAV_LINKS = [
